@@ -1,11 +1,11 @@
 import { Component } from '@angular/core';
 
 @Component({
-  selector: 'app-about',
-  standalone: true,
-  imports: [],
-  templateUrl: './about.component.html',
-  styleUrl: './about.component.sass'
+    selector: 'app-about',
+    standalone: true,
+    imports: [],
+    templateUrl: './about.component.html',
+    styleUrl: './about.component.sass'
 })
 export class AboutComponent {
 
