@@ -2,9 +2,10 @@ import { Component } from '@angular/core';
 import {AppConfig} from "../../config/app.config";
 
 @Component({
-  selector: 'app-settings',
-  templateUrl: './settings.component.html',
-  styleUrls: ['./settings.component.sass']
+    selector: 'app-settings',
+    templateUrl: './settings.component.html',
+    styleUrls: ['./settings.component.sass'],
+    standalone: false
 })
 export class SettingsComponent {
 
